Breaking out report types
