+++ app/src/js/components/ReconciliationReports/report-heading.js
@@ -17,6 +17,7 @@ const ReportHeading = ({
   onDownloadClick,
   reportState,
   startTime,
+  type,
 }) => {
   const breadcrumbConfig = [
     {
@@ -55,7 +56,7 @@ const ReportHeading = ({
         <div className="page__section__header">
           <div>
             <h1 className="heading--large heading--shared-content with-description ">
-              {name}
+              {type && `${type}: `}{name}
             </h1>
           </div>
           <div className="status--process">
@@ -127,6 +128,7 @@ ReportHeading.propTypes = {
   onDownloadClick: PropTypes.func,
   reportState: PropTypes.string,
   startTime: PropTypes.string,
+  type: PropTypes.string,
 };
 
 export default ReportHeading;
